+++ drizzle/0018_dazzling_the_hood.sql
@@ -1,0 +1,2 @@
+ALTER TABLE "domains" ADD COLUMN "dmarc_rua_email" varchar(255);--> statement-breakpoint
+ALTER TABLE "domains" ADD COLUMN "send_rate_per_minute" integer;